+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Formatted as described on http://keepachangelog.com/.
 
 - Test coverage (#25)
 - Timeout option (#21)
+- Workflow test for residues node (#24)
 
 ## [1.0.15] - 2016-10-18